feat(user): 유저서비스 인터페이스 삭제
- 인터페이스 삭제

67
