+++ src/main/java/com/account/yomankum/init/UserInitializer.java
@@ -4,12 +4,10 @@ import com.account.yomankum.user.domain.Role;
 import com.account.yomankum.user.domain.User;
 import com.account.yomankum.user.domain.UserType;
 import com.account.yomankum.user.domain.type.RoleName;
-import com.account.yomankum.user.dto.UserDto.UserSignUpDto;
 import com.account.yomankum.user.repository.UserRepository;
-import com.account.yomankum.user.service.UserService;
 import jakarta.annotation.PostConstruct;
 import java.time.Instant;
-import java.util.List;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Component;
